Update the way user level is read in from JSON to fix a bug with loading users. Story: 617

+++ src/main/java/seng302/Users/User.java
@@ -216,7 +216,7 @@ public class User {
 
         //Level
         try {
-            level = (int) properties.get("level");
+            level = ((Long) properties.get("level")).intValue();
         } catch (NullPointerException e) {
             //If level has never been set, (ie old account), default to 1
             level = 1;
